fix: emit reportTypeError template instantiations for all callers
reportTypeError / reportTypeErrorWithSuggestions were defined as variadic
templates in IRGenerator.cpp. Instantiations used only by other translation
units (HOFCodegen.cpp, OptionCombinatorsCodegen.cpp) — notably the
single-std::string_view overload — were therefore never emitted, leaving
undefined references at link time. The default Debug build happened to link
only because no optimization removed the in-TU uses; a Release build (with or
without tests) failed to link endo-bin.

Move both template definitions into IRGenerator.hpp (after the class) so every
calling TU can instantiate them, and add <format> for std::format_string.
